+++ NEWS.md
@@ -1,6 +1,6 @@
 # projector 0.1.0.9000
 
-* First release
+* First release, can plot matrix of vectots with T-SNE or PCA